Create separate types folder and barrel file for each type
