+++ src/types/chapter.ts
@@ -1,0 +1,5 @@
+export type Chapter = {
+  chapter: number;
+  title: string;
+  src: string;
+};
