+++ src/types/index.ts
@@ -1,0 +1,5 @@
+import { type Chapter } from "./chapter";
+import { type Subject } from "./subject";
+import { type Unit } from "./unit";
+
+export type { Chapter, Subject, Unit };
