+++ src/types/subject.ts
@@ -1,0 +1,6 @@
+import { type Unit } from ".";
+
+export type Subject = {
+  title: string;
+  units: Unit[];
+};
