+++ src/types/unit.ts
@@ -1,0 +1,7 @@
+import { type Chapter } from ".";
+
+export type Unit = {
+  unit: number;
+  title: string;
+  chapters: Chapter[];
+};
